Realize: You are dogshit at styling
You need a refresher on styling things
because it isn't acceptable that you are
incapable of basic flexbox maneuvers.
Googling having to justify-content and
slamming a footer to the bottom of a page
really isn't becoming of a pro developer.
Get the rust off your CSS skills and come back!